renting/requirements.txt
aiohttp==3.5.4
amqp==1.4.9
anyjson==0.3.3
APScheduler==3.6.0
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
Automat==0.7.0
backcall==0.1.0
beautifulsoup4==4.7.1
billiard==3.3.0.23
bleach==3.1.0
bs4==0.0.1
celery==3.1.26.post2
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
colorama==0.4.1
constantly==15.1.0
cryptography==2.6.1
cssselect==1.0.3
decorator==4.4.0
defusedxml==0.6.0
django==1.10.8
django-celery==3.2.2
entrypoints==0.3
et-xmlfile==1.0.1
fake-useragent==0.1.11
Flask==1.0.2
fonttools==3.44.0
hyperlink==19.0.0
idna==2.8
idna-ssl==1.1.0
incremental==17.5.0
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.13.3
Jinja2==2.10
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kombu==3.0.37
lxml==4.3.3
MarkupSafe==1.1.1
mistune==0.8.4
multidict==4.5.2
mysql-connector==2.2.9
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.15.0
openpyxl==2.6.2
pandas==0.23.3
pandocfilters==1.4.2
parsel==1.5.1
parso==0.4.0
pickleshare==0.7.5
pika==1.0.1
Pillow==5.3.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyDispatcher==2.0.5
PyExecJS==1.5.1
Pygments==2.4.0
PyHamcrest==1.9.0
pymongo==3.7.2
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pypinyin==0.35.2
pyquery==1.4.0
pyrsistent==0.15.2
python-dateutil==2.7.3
pytz==2019.1
pywin32==224
pywinpty==0.5.5
pyzmq==18.0.1
qtconsole==4.4.4
queuelib==1.5.0
rabbitmq==0.2.0
redis==3.2.1
requests==2.22.0
Scrapy==1.6.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.12.0
soupsieve==1.8
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
Twisted==19.2.0
typing-extensions==3.7.2
tzlocal==1.5.1
urllib3==1.25.3
vine==1.1.4
w3lib==1.20.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
yarl==1.3.0
zope.interface==4.6.0
